Remove core/configValidators.
Move those configs into Data Collector.
Add imsOrgId.
