--- src/core/configValidators.js
@@ -1,26 +1,0 @@
-import {
-  required,
-  validDomain,
-  eitherNilOrNonEmpty,
-  boolean
-} from "../utils/config-validators";
-
-export default {
-  propertyId: {
-    validate: required
-  },
-  edgeDomain: {
-    validate: validDomain,
-    defaultValue: "alpha.konductor.adobedc.net"
-  },
-  prehidingStyle: {
-    validate: eitherNilOrNonEmpty
-  },
-  authoringModeEnabled: {
-    defaultValue: false,
-    validate: boolean
-  },
-  // TODO: For debugging purposes only. Remove eventually.
-  shouldStoreCollectedData: { defaultValue: 1 },
-  device: { defaultValue: "Chrome-Mac" }
-};
